requirements.txt
numpy
torch
gitpython 
submitit
pandas
matplotlib
python-git